tests/wpt/tests/navigation-api/navigate-event/intercept-same-document-history-back.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  let start_length = navigation.entries().length;
  let start_index = navigation.currentEntry.index;

  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    let onnavigate_calls = 0;
    let onnavigatesuccess_calls = 0;
    navigation.onnavigate = e => {
      onnavigate_calls++;
      e.intercept();
    }
    navigation.onnavigatesuccess = t.step_func(e => {
      onnavigatesuccess_calls++;
      if (onnavigatesuccess_calls == 3) {
        assert_equals(navigation.entries().length, start_length + 2);
        assert_equals(navigation.currentEntry.index, start_index + 1);
        assert_equals(onnavigate_calls, 3);
        history.back();
      } else if (onnavigatesuccess_calls == 4) {
        assert_equals(navigation.entries().length, start_length + 2);
        assert_equals(navigation.currentEntry.index, start_index);
        assert_equals(onnavigate_calls, 4);
        t.done();
      }
    });

    navigation.navigate("?foo").finished
      .then(t.step_func(() => navigation.navigate("?bar").finished))
      .then(t.step_func(() => {
        assert_equals(navigation.entries().length, start_length + 2);
        assert_equals(navigation.currentEntry.index, start_index + 2);
        assert_equals(onnavigate_calls, 2);
        history.back();
      }));
  }, 0);
}, "event.intercept() can intercept same-document history.back()");
</script>
