LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-same-document-history-back-expected.txt


FAIL event.intercept() can intercept same-document history.back() assert_equals: expected 1 but got 2
